refactor: create post route fix (#65)
