+++ src/constants/queryKey.ts
@@ -1,3 +1,4 @@
 export const QUERY_KEYS = {
   STUDY_LIST: "studyList",
+  STUDY: "study",
 };
